refactored queries for readability; moved transformer out of engagement.service; moved repository parent classes and started using nestjs DI; made list call more efficient by reducing db calls;
