+++ src/components/engagement/engagement.service.ts
@@ -21,12 +21,7 @@ import {
   matchSession,
   property,
 } from '../../core';
-import {
-  parseBaseNodeProperties,
-  parsePropList,
-  parseSecuredProperties,
-  runListQuery,
-} from '../../core/database/results';
+import { hasMore } from '../../core/database/results';
 import { AuthorizationService } from '../authorization/authorization.service';
 import { CeremonyService } from '../ceremony';
 import { FileService } from '../file';
@@ -48,6 +43,7 @@ import {
   UpdateInternshipEngagement,
   UpdateLanguageEngagement,
 } from './dto';
+import { transformEngagementFromRepositoryToDTO } from './engagement.transformers';
 import {
   EngagementCreatedEvent,
   EngagementDeletedEvent,
@@ -58,39 +54,6 @@ import { EngagementRepository } from './repository/engagement.repository';
 
 @Injectable()
 export class EngagementService {
-  private readonly securedProperties = {
-    status: true,
-    statusModifiedAt: true,
-    completeDate: true,
-    disbursementCompleteDate: true,
-    communicationsCompleteDate: true,
-    initialEndDate: true,
-    startDate: true,
-    endDate: true,
-    startDateOverride: true,
-    endDateOverride: true,
-    modifiedAt: true,
-    lastSuspendedAt: true,
-    lastReactivatedAt: true,
-    ceremony: true,
-
-    //Language Specific
-    firstScripture: true,
-    lukePartnership: true,
-    sentPrintingDate: true,
-    paraTextRegistryId: true,
-    pnp: true,
-    language: true,
-
-    //Internship Specific
-    position: true,
-    growthPlan: true,
-    methodologies: true,
-    intern: true,
-    mentor: true,
-    countryOfOrigin: true,
-  };
-
   constructor(
     private readonly db: DatabaseService,
     private readonly ceremonyService: CeremonyService,
@@ -101,78 +64,10 @@ export class EngagementService {
     private readonly projectService: ProjectService,
     private readonly eventBus: IEventBus,
     private readonly authorizationService: AuthorizationService,
+    private readonly engagementRepository: EngagementRepository,
     @Logger(`engagement.service`) private readonly logger: ILogger
   ) {}
 
-  protected async transformEngagementsFromRepositoryToDTO(
-    result: any,
-    session: ISession
-  ) {
-    const props = parsePropList(result.propList);
-    const securedProperties = parseSecuredProperties(
-      props,
-      result.permList,
-      this.securedProperties
-    );
-
-    const project = await this.projectService.readOne(
-      result.projectId,
-      session
-    );
-
-    const canReadStartDate =
-      project.mouStart.canRead && securedProperties.startDateOverride.canRead;
-    const startDate = canReadStartDate
-      ? props.startDateOverride ?? project.mouStart.value
-      : null;
-    const canReadEndDate =
-      project.mouEnd.canRead && securedProperties.endDateOverride.canRead;
-    const endDate = canReadEndDate
-      ? props.endDateOverride ?? project.mouEnd.value
-      : null;
-
-    return {
-      __typename: result.__typename,
-      ...securedProperties,
-      ...parseBaseNodeProperties(result.node),
-      status: props.status,
-      modifiedAt: props.modifiedAt,
-      startDate: {
-        value: startDate,
-        canRead: canReadStartDate,
-        canEdit: false,
-      },
-      endDate: {
-        value: endDate,
-        canRead: canReadEndDate,
-        canEdit: false,
-      },
-      methodologies: {
-        ...securedProperties.methodologies,
-        value: securedProperties.methodologies.value ?? [],
-      },
-      ceremony: {
-        ...securedProperties.ceremony,
-        value: result.ceremonyId,
-      },
-      language: {
-        ...securedProperties.language,
-        value: result.languageId,
-      },
-      intern: {
-        ...securedProperties.intern,
-        value: result.internId,
-      },
-      countryOfOrigin: {
-        ...securedProperties.countryOfOrigin,
-        value: result.countryOfOriginId,
-      },
-      mentor: {
-        ...securedProperties.mentor,
-        value: result.mentorId,
-      },
-    };
-  }
   // CREATE /////////////////////////////////////////////////////////
 
   async createLanguageEngagement(
@@ -623,23 +518,21 @@ export class EngagementService {
     session: ISession
   ): Promise<LanguageEngagement | InternshipEngagement> {
     this.logger.debug('readOne', { id, userId: session.userId });
+    // TODO: this would work better early in the request handling pipeline.
+    //       Call could set session.userId to anon user if it is null.
+    //       It could always be assumed to exist and remove the duplicate logic.
+    if (!session.userId) {
+      this.logger.debug('using anon user id');
+      session.userId = this.config.anonUser.id;
+    }
 
-    const engagementRepository = new EngagementRepository(
-      this.db,
-      session,
-      this.files
-    );
-
-    const response = await engagementRepository
+    const response = await this.engagementRepository
       .request()
       .with({ engagementIds: [id] })
-      .hydrateEngagements(session)
-      .runQuery();
+      .hydrateEngagements(session.userId)
+      .runQuery(['engagements']);
 
-    return await this.transformEngagementsFromRepositoryToDTO(
-      response.engagements[0],
-      session
-    );
+    return transformEngagementFromRepositoryToDTO(response.engagements[0]);
   }
 
   // UPDATE ////////////////////////////////////////////////////////
@@ -898,23 +791,25 @@ export class EngagementService {
     { filter, ...input }: EngagementListInput,
     session: ISession
   ): Promise<EngagementListOutput> {
-    const engagementRepository = new EngagementRepository(
-      this.db,
-      session,
-      this.files
-    );
+    // TODO: this would work better early in the request handling pipeline.
+    //       Call could set session.userId to anon user if it is null.
+    //       It could always be assumed to exist and remove the duplicate logic.
+    if (!session.userId) {
+      this.logger.debug('using anon user id');
+      session.userId = this.config.anonUser.id;
+    }
 
-    const request = engagementRepository
+    const response = await this.engagementRepository
       .request()
-      .findEngagementIdsByProjectId(session, filter, input);
+      .findEngagementIdsByProjectId(session, filter, input)
+      .hydrateEngagements(session.userId)
+      .runQuery(['total', 'engagements']);
 
-    return await runListQuery(
-      request.query
-        .return(['items', 'total'])
-        .asResult<{ items: string[]; total: number }>(), // TODO: refactor, this is a bit obtuse
-      input,
-      (id) => this.readOne(id, session)
-    );
+    return {
+      total: response.total,
+      hasMore: hasMore(input, response.total),
+      items: response.engagements.map(transformEngagementFromRepositoryToDTO),
+    };
   }
 
   async listProducts(
