+++ src/core/database/query/lists.ts
@@ -23,23 +23,3 @@ export function calculateTotalAndPaginateList(
 
   return query.asResult<{ items: string[]; total: number }>();
 }
-
-export function calculateTotalAndPaginateListNoDBReturn(
-  query: Query,
-  { page, count, sort: sortInput, order }: SortablePaginationInput,
-  sort: Sorter
-) {
-  query
-    .with(['collect(distinct node) as nodes', 'count(distinct node) as total'])
-    .raw(`unwind nodes as node`)
-    // .with(['node', 'total']) TODO needed?
-    .call(sort, sortInput, order)
-    .with([
-      `collect(distinct node.id)[${(page - 1) * count}..${
-        page * count
-      }] as items`,
-      'total',
-    ]);
-
-  return query;
-}
